split guitar drawer into classes, fixed pom
