+++ .settings/org.eclipse.core.resources.prefs
@@ -1,7 +1,7 @@
 eclipse.preferences.version=1
 encoding//src/main/java=UTF-8
 encoding//src/main/java/log/charter/gui/CharterFrame.java=UTF-8
-encoding//src/main/java/log/charter/gui/chartPanelDrawers/instruments/GuitarDrawer.java=UTF-8
+encoding//src/main/java/log/charter/gui/chartPanelDrawers/instruments/guitar/GuitarDrawer.java=UTF-8
 encoding//src/main/java/log/charter/gui/components/ChordNameAdviceButton.java=UTF-8
 encoding//src/main/java/log/charter/util/chordRecognition/ChordNameAdder.java=UTF-8
 encoding//src/main/java/log/charter/util/chordRecognition/ChordNameSuggester.java=UTF-8
