+++ languages/English.txt
@@ -13,15 +13,15 @@ CONFIG_MINIMAL_TAIL_TO_NOTE_DISTANCE=Minimal distance between tail and next note
 CONFIG_MINIMAL_TAIL_LENGTH=Minimal note tail length
 CONFIG_SOUND_DELAY=Sound delay
 CONFIG_MARKER_POSITION=Marker position
+CONFIG_THEME=Theme
+CONFIG_THEME_DEFAULT=Default
+CONFIG_THEME_ROCKSMITH=Rocksmith
 CONFIG_NOTE_WIDTH=Note width
 CONFIG_NOTE_HEIGHT=Note height
 CONFIG_INVERT_STRINGS=Invert strings
 CONFIG_SHOW_CHORD_IDS=Show chord ids
 CONFIG_CREATE_DEFAULT_STRETCHES_IN_BACKGROUND=Create default stretches in background
 CONFIG_FPS=FPS
-CONFIG_THEME=Theme
-CONFIG_THEME_DEFAULT=Default
-CONFIG_THEME_ROCKSMITH=Rocksmith
 STRETCH_PANE=Custom music stretch
 STRETCH_PANE_VALUE=Song speed (%)
 EDIT_MENU=Edit
