+++ pom.xml
@@ -13,6 +13,11 @@
 	</properties>
 	<dependencies>
 		<dependency>
+			<groupId>com.googlecode.soundlibs</groupId>
+			<artifactId>tritonus-share</artifactId>
+			<version>0.3.7.4</version>
+		</dependency>
+		<dependency>
 			<groupId>org.jogamp.gluegen</groupId>
 			<artifactId>gluegen-rt</artifactId>
 			<version>2.3.2</version>
@@ -26,23 +31,11 @@
 			<groupId>com.googlecode.soundlibs</groupId>
 			<artifactId>mp3spi</artifactId>
 			<version>1.9.5-1</version>
-			<exclusions>
-				<exclusion>
-					<groupId>com.googlecode.soundlibs</groupId>
-					<artifactId>tritonus-share</artifactId>
-				</exclusion>
-			</exclusions>
 		</dependency>
 		<dependency>
 			<groupId>com.googlecode.soundlibs</groupId>
 			<artifactId>vorbisspi</artifactId>
 			<version>1.0.3-1</version>
-			<exclusions>
-				<exclusion>
-					<groupId>com.googlecode.soundlibs</groupId>
-					<artifactId>tritonus-share</artifactId>
-				</exclusion>
-			</exclusions>
 		</dependency>
 		<dependency>
 			<groupId>com.thoughtworks.xstream</groupId>
@@ -55,6 +48,7 @@
 			<version>0.7</version>
 		</dependency>
 	</dependencies>
+	
 	<build>
 		<finalName>RS Charter</finalName>
 		<plugins>
@@ -219,13 +213,4 @@
 		</plugins>
 	</build>
 	
-	<dependencyManagement>
-		<dependencies>
-			<dependency>
-				<groupId>com.googlecode.soundlibs</groupId>
-				<artifactId>tritonus-share</artifactId>
-				<version>0.3.7-2</version>
-			</dependency>
-		</dependencies>
-	</dependencyManagement>
 </project>
